# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  http.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
...
  models.Order:
    properties:
      created_at:
        type: string
      id:
        type: string
      payment_method:
        type: string
      shipping_address:
        $ref: '#/definitions/models.ShippingAddress'
      status:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      order_id:
        type: string
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  models.Product:
    properties:
      category_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  models.ShippingAddress:
    properties:
      city:
        type: string
      country:
        type: string
      postalCode:
        type: string
      street:
        type: string
    type: object
...
  order.CreateOrderRequset:
    properties:
      items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      payment_method:
        enum:
        - card
        - cash
        type: string
      shipping_adress:
        $ref: '#/definitions/models.ShippingAddress'
      user_id:
        type: string
    required:
    - items
    - payment_method
    - shipping_adress
    - user_id
    type: object
  order.OrderListResponse:
    properties:
      order:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      total:
        type: integer
    type: object
  order.UpdateOrderRequest:
    properties:
      status:
        enum:
        - pending
        - paid
        - shipped
        - canceled
        - delivered
        type: string
    required:
    - status
    type: object
...
  product.ProductListResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для e-commerce платформи з управлінням продуктами, кошиком та замовленнями
  title: E-Commerce API
  version: "1.0"
paths:
...
  /admin/orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Змінює статус конкретного замовлення
      parameters:
      - description: ID замовлення
        in: path
        name: id
        required: true
        type: string
      - description: Новий статус замовлення
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/order.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid request body or ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Оновлення статусу замовлення (Admin)
      tags:
      - admin
...
  /categories:
    get:
      consumes:
      - application/json
      description: Повертає список всіх категорій продуктів
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Отримати список категорій
      tags:
      - products
  /orders:
    get:
      consumes:
      - application/json
      description: Повертає список замовлень користувача з можливістю фільтрації за
        статусом та пагінацією
      parameters:
      - description: Фільтр по статусу
        enum:
        - pending
        - paid
        - shipped
        - canceled
        - delivered
        in: query
        name: status
        type: string
      - default: 20
        description: Кількість елементів на сторінку
        in: query
        name: limit
        type: integer
      - default: 0
        description: Зміщення для пагінації
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.OrderListResponse'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: User not authorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отримати список замовлень
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Створює нове замовлення для авторизованого користувача
      parameters:
      - description: Дані замовлення
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order.CreateOrderRequset'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: User not authorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Створити замовлення
      tags:
      - orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Повертає детальну інформацію про замовлення за його унікальним
        ідентифікатором
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отримати замовлення за ID
      tags:
      - orders
  /orders/{id}/cancel:
    put:
      consumes:
      - application/json
      description: Скасовує замовлення за ID (якщо воно ще не доставлене)
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order canceled message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Cannot cancel order
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Скасувати замовлення
      tags:
      - orders
  /products:
    get:
      consumes:
      - application/json
      description: Повертає список продуктів з можливістю фільтрації за категорією,
        ціною, наявністю та пагінацією
      parameters:
      - description: Category ID (UUID)
        in: query
        name: category_id
        type: string
      - description: Мінімальна ціна
        in: query
        name: min_price
        type: number
      - description: Максимальна ціна
        in: query
        name: max_price
        type: number
      - description: Пошуковий запит
        in: query
        name: search
        type: string
      - description: Тільки товари в наявності
        in: query
        name: in_stock
        type: boolean
      - default: 20
        description: Кількість елементів на сторінку
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Зміщення для пагінації
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.ProductListResponse'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Отримати список продуктів
      tags:
      - products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Повертає детальну інформацію про продукт за його унікальним ідентифікатором
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Отримати продукт за ID
      tags:
      - products
  /products/search:
    get:
      consumes:
      - application/json
      description: Повертає список продуктів, що відповідають пошуковому запиту
      parameters:
      - description: Пошуковий запит
        in: query
        name: q
        required: true
        type: string
      - default: 20
        description: Кількість елементів на сторінку
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Зміщення для пагінації
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "400":
          description: Search query is required or invalid parameters
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Пошук продуктів
      tags:
      - products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"